fix jwt

+++ app/file_upload/routes.py
@@ -1,8 +1,13 @@
 import os
 
-from fastapi import APIRouter, File, UploadFile
+from fastapi import APIRouter, File, UploadFile, status
 from .file_upload import _store_file
 from ..chatbot.vector_store.vectorstore import upload_to_vectorstore
+from fastapi.responses import JSONResponse
+from fastapi import HTTPException, Depends
+from app.auth.dependencies import get_current_active_user
+from app.auth.schemas import User
+from typing import Annotated
 
 router = APIRouter(
     prefix="/files",
@@ -14,15 +19,30 @@ UPLOAD_DIR = os.path.join(current_dir,"..","chatbot", "vector_store", "docs_reso
 
 
 @router.post("/file_upload")
-def file_upload(file: UploadFile = File(...), ):
+def file_upload(current_user: Annotated[User, Depends(get_current_active_user)] ,
+    file: UploadFile = File(...), 
+                 
+):
+    
+    filename = file.filename
+    if not (filename.endswith(".txt") or filename.endswith(".pdf")):
+        raise HTTPException(status_code=400, detail="Wrong file format.")
+
     # store the uploaded file
     file_name = _store_file(file, UPLOAD_DIR)
-    
+   
     #add to vector store
     file_name_from_vector = upload_to_vectorstore(file_name)
         
     # remove the file from the directory
     os.remove(os.path.join(UPLOAD_DIR, file_name_from_vector))
     
-    return {"file_name": file_name_from_vector}
-    
+    # return {"file_name": file_name_from_vector}
+    return JSONResponse(
+        status_code=status.HTTP_201_CREATED,
+        content = {
+            "message": "File uploaded successfully",
+            "success": True,
+            "filename": file_name_from_vector,
+        },
+    )
